✅ Standardize test entry point for database package (Task 5)
Create suite_test.go with integration build tags and remove TestDatabase
function from postgres_test.go to follow CLAUDE.md test suite standards.

Changes:
- Add internal/infrastructure/database/suite_test.go with integration tags
- Remove TestDatabase function from postgres_test.go
- Remove unused testing import from postgres_test.go
- Suite name: "Database Package Suite"

This ensures consistent test suite organization across the codebase and
proper integration test isolation via build tags.
