+++ internal/infrastructure/qrcode/suite_test.go
@@ -1,0 +1,13 @@
+package qrcode_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestQRCode(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "QRCode Generator Suite")
+}
